[ConstraintSystem] NFC: Move `Constraint.h` to `include/swift/Sema`
